+++ core/src/main/java/org/kiwix/kiwixmobile/core/di/modules/DownloaderModule.kt
@@ -18,8 +18,6 @@
 package org.kiwix.kiwixmobile.core.di.modules
 
 import android.app.DownloadManager
-import android.app.NotificationManager
-import android.content.Context
 import dagger.Module
 import dagger.Provides
 import org.kiwix.kiwixmobile.core.dao.DownloadRoomDao
@@ -30,7 +28,6 @@ import org.kiwix.kiwixmobile.core.downloader.DownloaderImpl
 import org.kiwix.kiwixmobile.core.downloader.downloadManager.DownloadManagerBroadcastReceiver
 import org.kiwix.kiwixmobile.core.downloader.downloadManager.DownloadManagerMonitor
 import org.kiwix.kiwixmobile.core.downloader.downloadManager.DownloadManagerRequester
-import org.kiwix.kiwixmobile.core.downloader.downloadManager.DownloadNotificationManager
 import org.kiwix.kiwixmobile.core.utils.SharedPreferenceUtil
 import javax.inject.Singleton
 
@@ -69,11 +66,4 @@ object DownloaderModule {
   fun providesDownloadManagerBroadcastReceiver(
     callback: DownloadManagerBroadcastReceiver.Callback
   ): DownloadManagerBroadcastReceiver = DownloadManagerBroadcastReceiver(callback)
-
-  @Provides
-  @Singleton
-  fun providesDownloadNotificationManager(
-    context: Context,
-    notificationManager: NotificationManager
-  ): DownloadNotificationManager = DownloadNotificationManager(context, notificationManager)
 }
